# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Library Management API",
      "description": "A RESTful API for managing a collection of books in a library."
    },
    "host": "localhost:5000",
    "basePath": "/",
    "schemes": ["http"],
    "paths": {
      "/books": {
        "post": {
          "summary": "Add a new book",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "isbn": { "type": "string" },
                  "title": { "type": "string" },
                  "author": { "type": "string" },
                  "published_year": { "type": "integer" }
                },
                "required": ["isbn", "title", "author", "published_year"]
              }
            }
          ],
          "responses": {
            "201": { "description": "Book added successfully" },
            "400": { "description": "Validation error" }
          }
        },
        "get": {
          "summary": "List all books",
          "responses": {
            "200": {
              "description": "A list of books",
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "isbn": { "type": "string" },
                    "title": { "type": "string" },
                    "author": { "type": "string" },
                    "published_year": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "/books/search": {
        "get": {
          "summary": "Search for books",
          "parameters": [
            { "name": "author", "in": "query", "type": "string" },
            { "name": "published_year", "in": "query", "type": "integer" }
          ],
          "responses": {
            "200": {
              "description": "Filtered list of books",
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "isbn": { "type": "string" },
                    "title": { "type": "string" },
                    "author": { "type": "string" },
                    "published_year": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "/books/{isbn}": {
        "delete": {
          "summary": "Delete a book by ISBN",
          "parameters": [
            { "name": "isbn", "in": "path", "required": true, "type": "string" }
          ],
          "responses": {
            "200": { "description": "Book deleted successfully" }
          }
        },
        "put": {
          "summary": "Update a book by ISBN",
          "parameters": [
            { "name": "isbn", "in": "path", "required": true, "type": "string" },
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "author": { "type": "string" },
                  "published_year": { "type": "integer" }
                }
              }
            }
          ],
          "responses": {
            "200": { "description": "Book updated successfully" },
            "404": { "description": "Book not found" }
          }
        }
      }
    }
  }
  